Split network settings out of server settings tab
