+++ idarling/interface/tabs/__init__.py
@@ -1,5 +1,6 @@
 from .servers import _TabCfgServer
 from .general import _TabCfgGeneral
+from .network import _TabCfgNetwork
 
 
 def TabCfgServer(program, parent):
@@ -8,3 +9,7 @@ def TabCfgServer(program, parent):
 
 def TabCfgGeneral(program, parent):
     return _TabCfgGeneral(program, parent).get()
+
+
+def TabCfgNetwork(program, parent):
+    return _TabCfgNetwork(program, parent).get()
